Setings Page Side Navigation
Utilizing previous work in d8787cb4 for anchor navigation, navigate to the specific card in the settings page when clicked.

I like this model mainly because in Mobile, users don't need to keep opening the drawer to navigate -- they just need to scroll. This allows us to use the same design for Mobile and App.

# Conflicts:
#	ui/page/settings/view.jsx